+++ packages/api-contract/src/types.ts
@@ -4,8 +4,8 @@
 import type { ApiBase } from '@polkadot/api/base';
 import type { ApiTypes } from '@polkadot/api/types';
 import type { Text, u64 } from '@polkadot/types';
-import type { ContractExecResultResult, ContractSelector, StorageDeposit } from '@polkadot/types/interfaces';
-import type { Codec, ICompact, INumber, TypeDef } from '@polkadot/types/types';
+import type { ContractExecResultResult, ContractSelector, StorageDeposit, WeightV2 } from '@polkadot/types/interfaces';
+import type { Codec, TypeDef } from '@polkadot/types/types';
 import type { BN } from '@polkadot/util';
 import type { Abi } from '.';
 
@@ -80,11 +80,6 @@ export interface ContractOptions {
 
 export interface BlueprintOptions extends ContractOptions {
   salt?: Uint8Array | string | null;
-}
-
-export interface WeightV2 {
-  refTime: ICompact<INumber>;
-  proofSize: ICompact<INumber>;
 }
 
 export interface WeightAll {