+++ packages/types-known/src/spec/westend.ts
@@ -18,7 +18,7 @@ const sharedTypes = {
   ProxyType: {
     _enum: ['Any', 'NonTransfer', 'Staking', 'SudoBalances', 'IdentityJudgement', 'CancelProxy']
   },
-  Weight: 'u64'
+  Weight: 'WeightV1'
 };
 
 const addrAccountIdTypes = {
@@ -92,8 +92,17 @@ const versioned: OverrideVersionedType[] = [
   },
   {
     // metadata v14
-    minmax: [9106, undefined],
-    types: {}
+    minmax: [9106, 9299],
+    types: {
+      Weight: 'WeightV1'
+    }
+  },
+  {
+    // metadata v14
+    minmax: [9300, undefined],
+    types: {
+      Weight: 'WeightV2'
+    }
   }
 ];